reorganize/recreate data directory, and remove superfluous stuff
data is in the format that the current xway version can eat. it is not
complete and will have to be recreated and/or checked, but it is as good
as necessary.

in the future this data might even be implicitly generated from the full
schedules (these used by the simulator), which currently are neither needed
nor done yet.