deuglify xway and ship inline version of fastgraph
the graph library might end up being heavily modified to be rvp-specific
(so that it is faster), so i put it to the project.

oh and there *is* a need to optimize. ask it to produce paths
from Hamborn Rathaus to OB-Hauptbahnhof at an elimination level of 2, 3, 4.
it just takes too long