graph: start an almost-failed implementation of LPA*
It produces some results. It is perfectly valid that it generates some
routes differently because they're equivalent. For example, these two partial
routes are currently both stored with a length of 3:

 (903 ) Vierlinden -> Fasanenstraße -> Walsum Betriebshof -> Walsum Rathaus
 (SB40) Vierlinden                  ->                       Walsum Rathaus

A recent commit removes the middle two entries to make diffing easier.

Still, this implementation of LPA* here drops two routes. Not a big loss,
but it's not supposed to happen. It can mean that the implementation of the
algorithm is faulty and just appears to work. Therefore it shouldn't be used
without further testing. Not yet.