simulate: first version that actually simulates something!
there are lots of issues with the current code now

it is unreadable.

it generates a very large number of result routes if there are too many
parallel connections for a physroute. for physroutes that are served
by many lines (for example, Bahnstrecke Köln-Duisburg), this can generate
an extraordinarily large number of pointless routes. see also the next
point and the one after that.

it doesn't have an intelligent switching system. currently any chance to
switch trains is taken (in a copy). the obvious solution would be to
forbid entering a line twice but this is wrong. assume a bus line serving
A <-> B <-> C <-> ... <-> Y <-> Z, taking like 60 minutes for all of it.
A and Z are also connected by a fast S-Bahn. If you want to get from
B to Y, then you have to enter the bus line twice to be faster. if you
forbid that, then the s-bahn route will be abandoned, because it cannot
bring you to the goal as it requires you to enter the bus line a second
time... and your journey takes 54 minutes instead of 15 or so. -- a fix
to this problem could be combined with one for the following one:

it also currently doesn't allow generating routes on the fly. it will become
impossible to keep all the routes that might be used in memory all the time.
therefore a way to generate routes on demand must be found. (the point is
that the simulator code *requests* them on demand, this system can also be
implemented with completely precalculated, hardcoded data...)

a fix that kills the last two problems with one stone would be to only
consider the generated routes if they provide an advantage over staying
on train. one way would be to prevent the alternative from being used
at all, which delays switching until the latest possible moment. other
method would be to mark affected trips specially and do stuff with them.
or to pick up that "forbid re-entering a line" thing again.. it can be
forbidden if reentering line A in the A-B-A sequence costs more than
A-B-B or A-A-A. whatever..