simulate: generate less events
This is, in some way, a regression.  The original idea was to be able to
trace all actions done by the VP.  However, thinking again about it,
this is really pointless.  Who wants to know all of that stuff?  For
example, the intermediate stations can be found in the route information
for the route that has been used to depart from the current station.  It's
recorded anyway (and always was).

This change makes the code simpler, though.