simulate: remove some unneeded TODO's regarding duplicate station support
The current code makes some assumptions - that on some route, a station
is never visited twice.  There are two cases when this can happen:

1. Ring lines - They have the same start and endpoint.  This case has
been specifically tested and was found to be working.  This is good.

2. Unusual operation conditions - This happened at the end of the 1990's
in Dortmund, where the U42 line visited a station twice (once above
ground and once below).  This will likely not work correctly if the
station is entered at the second occurence in the route (it will
put you to the first one instead and kill the route because the
next stop doesn't equal to the expected stop.)

As the first case isn't a problem, reasons why the second one is
crap have to be found:

1. Should there really be such a situation again, split the line in the
middle and link them, stating that the same car continues on the other
route.

2. These paths will never be generated by the planner anyway, so
it would only affect manually generated input for the purpose of
bare simulation.

Thanks to derf for pointing this out.