architectural change, introduces "calculate"
An old idea is picked up again here.  I think I will now represent the
routes in a graph again, with time information in them.  Why not?
What has been done in the simulator so far is nothing else than
constructing parts of the network in memory and then bruteforcing
the solutions (the trips).  Dijkstra will probably perform better than
a brute force search, so why not implement it and make life a bit
easier.

Of course this has to be implemented a bit cleaner than it is now.
Specifically date information has to be added (although the approach
with local time turns out to be useful).  Also, again a new method to
generate the data for the graph will need to be invented.  Specifically
a not so stupid on-demand loader/generator will have to be written.

One part of that problem could be solved by still keeping the planner
(xway) around to let it do some work.  But its role will change: it
will no longer dictate the simulator/calculator what exact routes to
take, but which route data to actually load.  So in theory the planner
still defines the route, but the calculator could theoretically come up
with completely new paths.  We'll see.