fahrplandaten: new route data storage format
Advantage is that this format is very close to the official schedule
books, and could therefore be generated from that.

Each *column* (except the first one of course) describes one trip.  Times are
the minute component of the timestamp.  If the hour is prefixed as "$h.", it is
used.  In the future there will be a syntax to store different arrive/depart
times, for now the only specified time is used for both.  Empty fields
obviously mean there is no train here.  There should never be two discontinuous
routes in one column.