setup some modules for managing all kinds of data (Line.pm, and so)
RVP is currently made more modularized, not only so that code isn't
duplicated across the different stages, but also so that little
test programs can be written much more quickly.  Also, cleanliness,
and it's easier to document.  And so on.  Thanks Captain Obvious.

I think I will get back to the simulator thing again.  However, it
will now internally use no timezones, no DST fuckup, no fancy stuff,
just Unix Time, as if it was local time, and everything will be well.
Also, while modularizing RVP, the planner can be extended to provide
information to the simulator that reduces the number of decisions
(and you know, in RVP, a decision is cloning all state).